Move range stats management to its own file.
