+++ components/region_cache_memory_engine/src/background.rs
@@ -1,31 +1,18 @@
 // Copyright 2024 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::{
-    cmp,
-    collections::{BTreeMap, BTreeSet},
-    fmt::Display,
-    num::NonZeroUsize,
-    sync::{
-        atomic::{AtomicBool, AtomicUsize, Ordering},
-        Arc,
-    },
-    thread::JoinHandle,
-    time::Duration,
-};
+use std::{collections::BTreeSet, fmt::Display, sync::Arc, thread::JoinHandle, time::Duration};
 
 use bytes::Bytes;
 use crossbeam::{
     channel::{bounded, tick, Sender},
     epoch, select,
-    sync::ShardedLock,
 };
 use engine_rocks::{RocksEngine, RocksSnapshot};
 use engine_traits::{
     CacheRange, IterOptions, Iterable, Iterator, MiscExt, RangeHintService, SnapshotMiscExt,
     CF_DEFAULT, CF_WRITE, DATA_CFS,
 };
-use kvproto::metapb::Region;
-use parking_lot::{Mutex, RwLock};
+use parking_lot::RwLock;
 use pd_client::{PdClient, RpcClient};
 use raftstore::coprocessor::RegionInfoProvider;
 use slog_global::{error, info, warn};
@@ -51,6 +38,7 @@ use crate::{
         RANGE_LOAD_TIME_HISTOGRAM,
     },
     range_manager::LoadFailedReason,
+    range_stats::RangeStatsManager,
     region_label::{
         LabelRule, RegionLabelAddedCb, RegionLabelRulesManager, RegionLabelServiceBuilder,
     },
@@ -319,235 +307,6 @@ struct BackgroundRunnerCore {
     engine: Arc<RwLock<RangeCacheMemoryEngineCore>>,
     memory_controller: Arc<MemoryController>,
     range_stats_manager: Option<RangeStatsManager>,
-}
-
-#[derive(Clone)]
-struct RangeStatsManager {
-    num_regions: Arc<AtomicUsize>,
-    info_provider: Arc<dyn RegionInfoProvider>,
-    prev_top_regions: Arc<Mutex<BTreeMap<u64, Region>>>,
-    checking_top_regions: Arc<AtomicBool>,
-    region_loaded_at: Arc<ShardedLock<BTreeMap<u64, Instant>>>,
-}
-
-/// Do not evict a region if has been cached for less than this duration.
-pub const EVICT_MIN_DURATION: Duration = Duration::from_secs(60 * 3);
-
-impl RangeStatsManager {
-    /// Initial number of top regions to track and cache. This may change, see
-    /// `adjust_max_num_regions` below.
-    pub fn new(num_regions: usize, info_provider: Arc<dyn RegionInfoProvider>) -> Self {
-        RangeStatsManager {
-            num_regions: Arc::new(AtomicUsize::new(num_regions)),
-            info_provider,
-            prev_top_regions: Arc::new(Mutex::new(BTreeMap::new())),
-            checking_top_regions: Arc::new(AtomicBool::new(false)),
-            region_loaded_at: Arc::new(ShardedLock::new(BTreeMap::new())),
-        }
-    }
-
-    /// Prevents two instances of this from running concurrently.
-    pub fn set_checking_top_regions(&self, v: bool) {
-        self.checking_top_regions.store(v, Ordering::Relaxed);
-    }
-
-    /// Returns true if another thread is checking top regions.
-    pub fn checking_top_regions(&self) -> bool {
-        self.checking_top_regions.load(Ordering::Relaxed)
-    }
-
-    fn set_max_num_regions(&self, v: usize) {
-        self.num_regions.store(v, Ordering::Relaxed);
-    }
-
-    /// Returns the maximum number of regions that can be cached.
-    ///
-    /// See also `adjust_max_num_regions` below.
-    pub fn max_num_regions(&self) -> usize {
-        self.num_regions.load(Ordering::Relaxed)
-    }
-
-    /// Collect candidates for eviction sorted by activity in creasing order:
-    ///
-    /// 1. Get all the regions sorted (decreasing) by region activity using
-    ///    [raftstore::coprocessor::RegionCollector::handle_get_top_regions].
-    /// 2. Remove all regions where `is_cached_pred` returns false when passed
-    ///    the region's range or those that have been loaded for less than
-    ///    [`EVICT_MIN_DURATION`].
-    /// 3. Reverse the list so that it is now sorted in the order of increasing
-    ///    activity.
-    /// 4. Store the results in `ranges_out` using [Vec::extend].
-    pub fn collect_candidates_for_eviction<F>(
-        &self,
-        ranges_out: &mut Vec<(CacheRange, u64)>,
-        is_cached_pred: F,
-    ) where
-        F: Fn(&CacheRange) -> bool,
-    {
-        // Gets all of the regions, sorted by activity.
-        let all_regions = self.info_provider.get_top_regions(None).unwrap();
-        let regions_loaded = self.region_loaded_at.read().unwrap();
-        ranges_out.extend(
-            all_regions
-                .iter()
-                .filter_map(|(region, approx_size)| {
-                    let r = CacheRange::from_region(region);
-                    is_cached_pred(&r)
-                        .then(|| {
-                            match regions_loaded.get(&region.get_id()) {
-                                // Do not evict ranges that were loaded less than
-                                // `EVICT_MIN_DURATION` ago.
-                                Some(&time_loaded)
-                                    if Instant::now() - time_loaded < EVICT_MIN_DURATION =>
-                                {
-                                    None
-                                }
-                                Some(_) | None =>
-                                // None indicates range loaded from a hint, not by this manager.
-                                {
-                                    Some((r, *approx_size))
-                                }
-                            }
-                        })
-                        .flatten()
-                })
-                .rev(),
-        );
-    }
-
-    /// This method should be called when `evicted_range` is succesfully evicted
-    /// to remove any internal `RegionStatsManager` that corresponds to the
-    /// range.
-    ///
-    /// Calls [raftstore::coprocessor::region_info_accessor::RegionInfoProvider::find_region_by_key] to
-    /// find the region corresponding to the range.
-    ///
-    /// TODO (afeinberg): This is inefficient, either make this method bulk, or
-    /// find another way to avoid calling `find_region_by_key` in a loop.
-    pub fn handle_range_evicted(&self, evicted_range: &CacheRange) {
-        // TODO (afeinberg): This is inefficient.
-        let _ = self
-            .info_provider
-            .find_region_by_key(&evicted_range.start)
-            .map(|region| {
-                let id = region.get_id();
-                let _ = self.prev_top_regions.lock().remove(&id);
-                let _ = {
-                    let mut regions_loaded = self.region_loaded_at.write().unwrap();
-                    regions_loaded.remove(&id)
-                };
-            });
-    }
-
-    /// Attempt to adjust the maximum number of cached regions based on memory
-    /// usage:
-    ///
-    /// If `curr_memory_usage` is LESS THAN `threshold` by 3 *
-    /// [`EXPECTED_AVERAGE_REGION_SIZE`] bytes, *increase* the maximum
-    /// by `threshold - curr_memory_usage / 3 * EXPECTED_AVERAGE_REGION_SIZE`.
-    ///
-    /// If `curr_memory_usage` is GREATER THAN `threshold`, *decrease* the
-    /// maximum by `(curr_memory_usage - threshold) /
-    /// EXPECTED_AVERAGE_REGION_SIZE`.
-    pub fn adjust_max_num_regions(&self, curr_memory_usage: usize, threshold: usize) {
-        match curr_memory_usage.cmp(&threshold) {
-            cmp::Ordering::Less => {
-                let room_to_grow = threshold - curr_memory_usage;
-                if room_to_grow > EXPECTED_AVERAGE_REGION_SIZE * 3 {
-                    let curr_num_regions = self.max_num_regions();
-                    let next_num_regions =
-                        curr_num_regions + room_to_grow / (EXPECTED_AVERAGE_REGION_SIZE * 3);
-                    info!("increasing number of top regions to cache";
-                        "from" => curr_num_regions,
-                        "to" => next_num_regions,
-                    );
-                    self.set_max_num_regions(next_num_regions);
-                }
-            }
-            cmp::Ordering::Greater => {
-                let to_shrink_by = curr_memory_usage - threshold;
-                let curr_num_regions = self.max_num_regions();
-                let next_num_regions = curr_num_regions
-                    .checked_sub(1.max(to_shrink_by / EXPECTED_AVERAGE_REGION_SIZE))
-                    .unwrap_or(1)
-                    .max(1);
-                info!("decreasing number of top regions to cache";
-                    "from" => curr_num_regions,
-                    "to" => next_num_regions,
-                );
-                self.set_max_num_regions(next_num_regions);
-            }
-            _ => (),
-        };
-    }
-
-    /// Collects changes to top regions since the previous time this method was
-    /// called. This method should be called by background tasks responsing
-    /// for algorithmic loading and eviction.
-    ///
-    /// 1. Calls [raftstore::coprocessor::RegionCollector::handle_get_top_regions] to
-    ///    request the top `self.max_num_regions()` regions.
-    ///
-    /// 2. If this is the first time this method has been called on this
-    ///    instance, stores the results of previous step in `ranges_added_out`
-    ///    and returns.
-    ///
-    /// 3. If this method has been called before, compare results of step 1 with
-    ///    previous results:
-    ///   - Newly added ranges (regions missing from previous results) are
-    ///     stored in `ranges_added_out`. This can happen when
-    ///     `max_num_regions()` increases, or when `max_num_regions()` is
-    ///     unchanged but the activity order changed.
-    ///   - Removed regions - regions included in previous results - but not the
-    ///     current ones are stored in `ranges_removed_out`.
-    pub fn collect_changed_ranges(
-        &self,
-        ranges_added_out: &mut Vec<CacheRange>,
-        ranges_removed_out: &mut Vec<CacheRange>,
-    ) {
-        info!("collect_changed_ranges"; "num_regions" => self.max_num_regions());
-        let curr_top_regions = self
-            .info_provider
-            .get_top_regions(Some(NonZeroUsize::try_from(self.max_num_regions()).unwrap()))
-            .unwrap() // TODO (afeinberg): Potentially custom error handling here.
-            .iter()
-            .map(|(r, _)| (r.id, r.clone()))
-            .collect::<BTreeMap<_, _>>();
-        {
-            let mut region_loaded_map = self.region_loaded_at.write().unwrap();
-            for &region_id in curr_top_regions.keys() {
-                let _ = region_loaded_map.insert(region_id, Instant::now());
-            }
-        }
-        let prev_top_regions = {
-            let mut mut_prev_top_regions = self.prev_top_regions.lock();
-            let ret = mut_prev_top_regions.clone();
-            *mut_prev_top_regions = curr_top_regions.clone();
-            ret
-        };
-        if prev_top_regions.is_empty() {
-            ranges_added_out.extend(curr_top_regions.values().map(CacheRange::from_region));
-            return;
-        }
-        let added_ranges = curr_top_regions
-            .iter()
-            .filter(|(id, _)| !prev_top_regions.contains_key(id))
-            .map(|(_, region)| CacheRange::from_region(region));
-        let regions_loaded = self.region_loaded_at.read().unwrap();
-        let removed_ranges = prev_top_regions.iter().filter_map(|(id, region)| {
-            if !curr_top_regions.contains_key(id) {
-                match regions_loaded.get(id) {
-                    // Do not evict ranges that were loaded less than `EVICT_MIN_DURATION` ago.
-                    Some(&time_loaded) if Instant::now() - time_loaded < EVICT_MIN_DURATION => None,
-                    _ => Some(CacheRange::from_region(region)),
-                }
-            } else {
-                None
-            }
-        });
-        ranges_added_out.extend(added_ranges);
-        ranges_removed_out.extend(removed_ranges);
-    }
 }
 
 impl BackgroundRunnerCore {
