+++ components/region_cache_memory_engine/src/lib.rs
@@ -22,6 +22,7 @@ mod memory_controller;
 mod metrics;
 mod perf_context;
 mod range_manager;
+mod range_stats;
 mod read;
 mod region_label;
 mod statistics;
